requirements.txt
# =====================================
# Web Framework and Extensions
# =====================================
fastapi[all]
slowapi

# =====================================
# Data Validation and Serialization
# =====================================
pydantic

# =====================================
# Web Server
# =====================================
uvicorn

# =====================================
# Configuration Management
# =====================================
python-decouple

# =====================================
# Database and ORM
# =====================================
## sqlalchemy
## databases

# =====================================
# Authentication and Authorization
# =====================================
## passlib[bcrypt]
## jwt


# =====================================
# Asynchronous Libraries
# =====================================
httpx
## aioredis

# =====================================
# Machine Learning and AI
# =====================================
transformers
torch

# =====================================
# Utilities
# =====================================
requests

# =====================================
# Testing
# =====================================
pytest
pytest-asyncio
pytest-mock
selenium
